Created/updated the ANOVA/ANCOVA macro, with loops and outputs
Originally, the ANOVA macro used two macros, with one imbedded into
the other.  It was flexible and had to have several marco variables
called externally, so it wasn't modular.  It also didn't output
anything.  I have updated it to loop through the specified variables
(making it more flexible), no external variables need to be called,
and it outputs data into a format useful for pushing directly into a
table format in a report, manuscript, or presentation (poster or
oral).