Moved the gee macro into its own file, added doc, updated doc
I moved the gee macro into its own file as I am slowly moving my
macros so that they are as a SAS autocall library.  I renamed the
For.sas file to for.sas as the autocall library wasn't picking up the
uppercase F.  Added documentation to the gee.sas file.  Updated the
documentation.