Updated README, merged readme to work with pdf + github
A few commits back I created two READMEs, one for github and one for
pdf. However, that is not ideal at all. So I combined them, moved the
pdf YAML header into a hidden file, and created a Makefile to make the
pdf properly while keeping the md in the GitHub format.